Refactor chatinput functionality
