+++ app/components/coding-exercise/ui/chat-panel-states/StuckHeader.tsx
@@ -1,17 +1,13 @@
 import ChatBubbleIcon from "@/icons/chat-bubble.svg";
 import styles from "./FreeUserCanStart.module.css";
 
-interface StuckHeaderProps {
-  title?: string;
-}
-
-export default function StuckHeader({ title = "Feeling Stuck? Talk to Jiki!" }: StuckHeaderProps) {
+export default function StuckHeader() {
   return (
     <>
       <div className={styles.avatar}>
         <ChatBubbleIcon width={32} height={32} />
       </div>
-      <h3 className={styles.title}>{title}</h3>
+      <h3 className={styles.title}>Feeling Stuck? Talk to Jiki!</h3>
     </>
   );
 }
